filter|msresamp: fleshing out interp method

+++ include/liquid.h
@@ -1639,7 +1639,7 @@ void MSRESAMP(_execute)(MSRESAMP() _q,                              \
                         unsigned int _nx,                           \
                         TO * _y,                                    \
                         unsigned int *_ny);                         \
-unsigned int MSRESAMP(_get_delay)(MSRESAMP() _q);                   \
+float MSRESAMP(_get_delay)(MSRESAMP() _q);                          \
 
 LIQUID_MSRESAMP_DEFINE_API(MSRESAMP_MANGLE_RRRF,
                            float,
